requirements.txt
streamlit==1.32.2
librosa==0.10.1
numpy==1.24.4
matplotlib==3.7.5
torch==2.2.2
torchaudio==2.2.2
openai-whisper==20231117
scikit-learn==1.3.2
scipy==1.11.4
soundfile==0.12.1
